Move utils.linalg -> preprocessing.color_deconvolution._linalg
