+++ histomicstk/preprocessing/color_deconvolution/_linalg.py
